feat(main): Update OnlinePay - Add payment method to create schedule - Add payment status to response - Change response in redis repository from Integer to Object - Implement extract pay

+++ booking-server-command/src/main/java/com/example/bookingserver/application/command/handle/user/VerifyOTPHandler.java
@@ -17,7 +17,7 @@ public class VerifyOTPHandler implements Handler<VerifyOTPCommand> {
     @Override
     @SneakyThrows
     public void execute(VerifyOTPCommand command) {
-        Integer codeForOncePerson=redisRepository.get(command.getEmail());
+        Integer codeForOncePerson= (Integer) redisRepository.get(command.getEmail());
         if(codeForOncePerson == null || !codeForOncePerson.equals(command.getCode())){
             throw new BookingCareException(ErrorDetail.ERR_USER_UN_AUTHENTICATE);
         }
